+++ playwright.config.ts
@@ -12,8 +12,10 @@ export default defineConfig({
       name: 'library',
       testMatch: [
         '**/anvil.spec.ts',
+        '**/fixtures-multichain.spec.ts',
         '**/live-fixtures.spec.ts',
         '**/mock-wallet.spec.ts',
+        '**/mock-wallet-multichain.spec.ts',
         '**/private-key-rpc-client.spec.ts',
         '**/provider-injection.spec.ts',
         '**/real-wallet.spec.ts',